refactored the sub components of CardModal and changed autoHeightTextarea folder name to AutoHeightTextarea
