+++ client/src/components/CardModal/ActivityComponent/index.js
@@ -1,0 +1,3 @@
+import ActivityComponent from './ActivityComponent';
+
+export default ActivityComponent;
