+++ client/src/components/CardModal/CommentBox/index.js
@@ -1,0 +1,3 @@
+import CommentBox from "./CommentBox";
+
+export default CommentBox;
