+++ client/src/components/CardModal/DescriptionComponent/index.js
@@ -1,0 +1,3 @@
+import DescriptionComponent from './DescriptionComponent';
+
+export default DescriptionComponent;
