+++ client/src/components/CardModal/SideActionsComponent/index.js
@@ -1,0 +1,3 @@
+import SideActionsComponent from './SideActionsComponent';
+
+export default SideActionsComponent;
