+++ client/src/components/CardModal/TitleComponent/TitleComponent.jsx
@@ -4,8 +4,8 @@ import { MdTitle } from "react-icons/md";
 import { AiOutlineEye } from "react-icons/ai";
 
 import classes from "./TitleComponent.module.css";
-import { CardListContext } from "../../context/cardlist-context";
-import AutoHeightTextarea from "../autoHeightTextarea";
+import { CardListContext } from "../../../context/cardlist-context";
+import AutoHeightTextarea from "../../AutoHeightTextarea";
 
 export default function TitleComponent({ cardId }) {
   const [isEditMode, setIsEditMode] = useState(false);
