+++ client/src/components/CardModal/TitleComponent/index.js
@@ -1,0 +1,3 @@
+import TitleComponent from './TitleComponent';
+
+export default TitleComponent        ;
